python/django/pyproject.toml
[project]
name = "django-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.16",
]
